Jun 7 2016 - Bug Fixes
Fixed a Bug in the Class creator function. an end-brace ("}") was being
placed in the wrong position, causing errors in the Public Search
functions.

Added the ability to set the naming conventions in the SQL Tables and
Stored Procedures.  In the ClassBuilder.cs file, there is a constant
called "SQL_NAMING_CONVENTION".  Set that constant to 1 or 2
(explanation of the naming conventions are in the code).